requirements.txt
bcrypt==4.2.1
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz==2024.2
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
gunicorn==20.1.0
whitenoise==6.5.0  # Added for serving static files
